Add request logging filter and update API routing
A new request logging filter is added to log all incoming requests for tracing and debugging purposes. Simultaneously, changes are made to the API routing to include '/api' in the path of the endpoints to standardize the URL pattern. Besides, modifications were done to swagger configurations, pom.xml, application properties and some changes in unit test assumptions.